requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic[email]==2.7.1
firebase-admin >=6.5.0
google-cloud-secret-manager
google-cloud-storage>=2.10.0
Pillow==10.3.0
python-multipart==0.0.9
httpx==0.27.0
google-cloud-kms==2.19.0
cryptography==42.0.8
PyJWT==2.8.0
h2==4.1.0